sync PDF page, refactoring
