+++ src/scene-entry-manager.js
@@ -1,6 +1,7 @@
 import qsTruthy from "./utils/qs_truthy";
 import screenfull from "screenfull";
 import nextTick from "./utils/next-tick";
+import pinnedEntityToGltf from "./utils/pinned-entity-to-gltf";
 
 const playerHeight = 1.6;
 const isBotMode = qsTruthy("bot");
@@ -213,31 +214,8 @@ export default class SceneEntryManager {
 
     this.scene.addEventListener("object_pinned", e => {
       const el = e.detail.el;
-      if (!NAF.utils.isMine(el)) return;
-
-      // Construct a GLTF node from this entity
-      const object3D = el.object3D;
-      const components = el.components;
-      const networkId = components.networked.data.networkId;
-
-      const gltfComponents = {};
-      const gltfNode = { name: networkId, extensions: { HUBS_components: gltfComponents } };
-
-      // Adapted from three.js GLTFExporter
-      const equalArray = (x, y) => x.length === y.length && x.every((v, i) => v === y[i]);
-      const rotation = object3D.quaternion.toArray();
-      const position = object3D.position.toArray();
-      const scale = object3D.scale.toArray();
-
-      if (!equalArray(rotation, [0, 0, 0, 1])) gltfNode.rotation = rotation;
-      if (!equalArray(position, [0, 0, 0])) gltfNode.translation = position;
-      if (!equalArray(scale, [1, 1, 1])) gltfNode.scale = scale;
-
-      if (components["media-loader"]) {
-        gltfComponents.media = { src: components["media-loader"].data.src, networkId };
-      }
-
-      gltfComponents.pinnable = { pinned: true };
+      const networkId = el.components.networked.data.networkId;
+      const gltfNode = pinnedEntityToGltf(el);
 
       this.hubChannel.pin(networkId, gltfNode);
     });
